貸付制度結果ページのリファクタリング。miiboのURL更新。
