+++ dashboard/src/config/app_config.json
@@ -6,7 +6,7 @@
     },
     "Github": "https://github.com/project-inclusive/OpenFisca-Japan",
     "contact": "https://www.poverty-prevention.net/",
-    "chatbot": "https://miibo.jp/chat/f791f806-10f2-493c-8ed5-0174ec3e483618afd9cdeff118?name=%E3%83%A4%E3%83%89%E3%82%AB%E3%83%AA%E3%81%8F%E3%82%93",
+    "chatbot": "https://miibo.jp/chat/b86882cc-35e6-4796-a880-b2148056299418ce92567e3f9?name=%E3%83%A4%E3%83%89%E3%82%AB%E3%83%AA%E3%81%8F%E3%82%93",
     "questionnaire_form": "https://forms.gle/fxztn7JGTuZ3v1mS9"
   },
   "style": {
